move the arrow extension type + test
