+++ CMakeLists.txt
@@ -261,7 +261,7 @@ if(GEOARROW_BUILD_TESTS)
   add_executable(array_reader_test src/geoarrow/array_reader_test.cc)
   add_executable(array_writer_test src/geoarrow/array_writer_test.cc)
   add_executable(wkx_files_test src/geoarrow/wkx_files_test.cc)
-  add_executable(geoarrow_arrow_test src/geoarrow/geoarrow_arrow_test.cc)
+  add_executable(hpp_arrow_extension_type_test src/geoarrow/hpp/arrow_extension_type_test.cc)
 
   target_link_libraries(geoarrow_hpp_test
                         geoarrow
@@ -285,7 +285,7 @@ if(GEOARROW_BUILD_TESTS)
   target_link_libraries(array_reader_test geoarrow gtest_main nanoarrow::nanoarrow)
   target_link_libraries(array_writer_test geoarrow gtest_main nanoarrow::nanoarrow)
   target_link_libraries(wkx_files_test geoarrow gtest_main nanoarrow::nanoarrow)
-  target_link_libraries(geoarrow_arrow_test geoarrow ${GEOARROW_ARROW_TARGET} gtest_main)
+  target_link_libraries(hpp_arrow_extension_type_test geoarrow ${GEOARROW_ARROW_TARGET} gtest_main)
 
   include(GoogleTest)
   gtest_discover_tests(geoarrow_hpp_test)
@@ -305,5 +305,5 @@ if(GEOARROW_BUILD_TESTS)
   gtest_discover_tests(array_reader_test)
   gtest_discover_tests(array_writer_test)
   gtest_discover_tests(wkx_files_test)
-  gtest_discover_tests(geoarrow_arrow_test)
+  gtest_discover_tests(hpp_arrow_extension_type_test)
 endif()
